+++ src/config.ts
@@ -1,14 +1,13 @@
 import { store } from "./persistent";
-import { Configuration } from "./types";
 
-export function set(configuration: Partial<Configuration>, options?: { lowPriority?: boolean }) {
-    const cfgA = options?.lowPriority ? configuration : store.configuration.$value;
-    const cfgB = options?.lowPriority ? store.configuration.$value : configuration;
-
-    store.configuration.value = {
-        ...cfgA,
-        ...cfgB,
-        decoration: { ...cfgA.decoration, ...cfgB.decoration },
-    };
+export function disableTags(tags: string[]) {
+    store.config.value.t0.push(...tags);
 }
 
+export function enableTags(tags: string[]) {
+    store.config.value.t1.push(...tags);
+}
+
+export function disableVariable() {
+    store.config.value.v0 = true;
+}