+++ src/persistent.ts
@@ -1,15 +1,12 @@
 import { persistentValue } from "persival";
-import { Configuration, Message } from "./types";
+import { Config, Message } from "./types";
 
 const root = "_logMess";
 
 const messages = persistentValue<Message[]>([], { name: "messages", root });
-const configuration = persistentValue<Configuration>(
-    { mode: "both", decoration: {} },
-    { name: "configuration", root }
-);
+const config = persistentValue<Config>({ t0: [], t1: [], v0: false }, { name: "config", root });
+const id = persistentValue<number>(0, { name: "id", root });
 
-const store = { messages, configuration };
+const store = { messages, config, id };
 
 export { store };
-