+++ src/tag.ts
@@ -1,20 +1,16 @@
-import { Decoration } from "./types";
-
-export function getTag(tag: string | undefined, decoration: Decoration): string[] {
+export function getTag(
+    tag: string | undefined,
+    decoration?: { bg?: string; fg?: string }
+): string[] {
     if (!tag) return [];
-
-    if (navigator.userAgent.indexOf("Trident") !== -1) {
-        return [`[${tag}]`];
-    }
+    if (navigator.userAgent.indexOf("Trident") !== -1) return [`[${tag}]`];
 
     const tagText = `%c ${tag} `;
 
-    if (decoration[tag])
+    if (decoration?.bg || decoration?.fg)
         return [
             tagText,
-            `background: ${decoration[tag].bgColor}; color: ${decoration[tag].fgColor};`,
+            `background: ${decoration.bg || "white"}; color: ${decoration.fg || "black"};`,
         ];
-
     return [tagText, "border: 1px solid black;"];
 }
-